Move .show of showResult to keyup instead of click
-Having the search results div shows the previous search's results on
clicking on the box. This will not show newly created creatures and
will show the creature deleted immediately before clicking on the
search field. It also will initially show a blank div that, when clicked
on, produces a null value.
-Moving the .show to the keyup part of the search function makes sure
all results that appear are up-to-date and prevents access to the null
div.